*** empty log message ***

+++ layer2/CoordSet.c
@@ -531,11 +531,11 @@ void CoordSetRender(CoordSet *I,CRay *ray,Pickable **pick)
           } else {
             if(I->Obj) 
               ray->fTexture(ray,
-                            SettingGet_f(I->Setting,I->Obj->Obj.Setting,cSetting_ray_texture),
+                            (int)SettingGet_f(I->Setting,I->Obj->Obj.Setting,cSetting_ray_texture),
                             SettingGet_fv(I->Setting,I->Obj->Obj.Setting,cSetting_ray_texture_settings));
             else
               ray->fTexture(ray,
-                            SettingGet_f(I->Setting,NULL,cSetting_ray_texture),
+                            (int)SettingGet_f(I->Setting,NULL,cSetting_ray_texture),
                             SettingGet_fv(I->Setting,NULL,cSetting_ray_texture_settings));
             ray->fColor3fv(ray,ColorGet(I->Obj->Obj.Color));
           }
